+++ diamondback/__init__.py
@@ -73,4 +73,4 @@ from .transforms import FourierTransform  # noqa: F401
 from .transforms import PowerSpectrumTransform  # noqa: F401
 from .transforms import WaveletTransform  # noqa: F401
 from .transforms import ZTransform  # noqa: F401
-from .versions import __version__
+from .__metadata__ import __version__  # noqa: F401