+++ diamondback/__metadata__.py
@@ -1,0 +1,19 @@
+""" **Description**
+
+        Metadata.
+
+    **License**
+
+        `BSD-3C. <https://github.com/larryturner/diamondback/blob/master/license>`_
+
+        © 2018 - 2021 Larry Turner, Schneider Electric Industries SAS. All rights reserved.
+
+    **Author**
+
+        Larry Turner, Schneider Electric, Analytics & AI, 2018-03-22.
+
+    **Definition**
+
+"""
+
+__version__ = '1.0.90'