+++ changelog.rst
@@ -473,3 +473,10 @@ Release
 ^^^^^^^^^^^^^^^^^^^
 
 -   Modified imports.
+
+1.0.90 - 2021-05-11
+^^^^^^^^^^^^^^^^^^^
+
+-   Added __metadata.py to import metadata without dependencies.
+
+-   Modified setup.